1.login.html 2.change password use JPA 3.add interceptor

+++ src/main/java/com/felix/shoppingcentre/repository/UserRepository.java
@@ -4,6 +4,6 @@ import com.felix.shoppingcentre.entity.User;
 import com.felix.shoppingcentre.entity.UserJpaEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserJpaEntity, Long> {
+public interface UserRepository extends CrudRepository<UserJpaEntity, Integer> {
     UserJpaEntity findByUsername(String username);
 }
